requirements.txt
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2019.11.28
chardet==3.0.4
click==8.1.7
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
flasgger==0.9.7.1
flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Login==0.6.2
greenlet==1.1.2
idna==2.8
importlib-metadata==6.8.0
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
mistune==3.0.1
mysqlclient==2.1.0
numpy==1.15.0
packaging==23.1
pep8==1.7.1
pkgutil-resolve-name==1.3.10
psutil==5.9.0
pycodestyle==2.8.0
PyGObject==3.36.0
python-apt==2.0.0+ubuntu0.20.4.7
python-dotenv==1.0.0
PyYAML==6.0.1
referencing==0.30.2
requests==2.22.0
requests-unixsocket==0.2.0
rpds-py==0.10.3
six==1.14.0
SQLAlchemy==1.4.32
ssh-import-id==5.10
unattended-upgrades==0.1
urllib3==1.25.8
werkzeug==2.3.7
zipp==3.17.0
